app/requirements.txt
# FastAPI 관련
fastapi==0.95.0
uvicorn==0.21.0
pydantic==1.10.0

# 데이터베이스
mysql-connector-python==8.0.32
redis==4.5.4
pymongo==4.3.3

# 인증 및 보안
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# 모니터링 및 분석
prometheus-client==0.16.0
elasticsearch==8.0.0

# 유틸리티
python-dotenv==1.0.0
requests==2.28.2
cryptography==40.0.0
